+++ client/test/protractor-conf.js
@@ -14,7 +14,7 @@ exports.config = {
         }
     },
 
-    'chromeOnly': true,
+    //'directConnect': true,
 
     'baseUrl': 'http://localhost:8000/',